Move the toolpad app out of components folder
