+++ packages/toolpad-app/pages/_toolpad/[[...index]].tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from 'next';
 import * as React from 'react';
-import Toolpad from '../../src/components/Toolpad';
+import Toolpad from '../../src/toolpad';
 
 const Home: NextPage = () => <Toolpad basename="/_toolpad" />;
 export default Home;
