+++ packages/toolpad-app/pages/index.tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from 'next';
 import * as React from 'react';
-import Home from '../src/components/Home';
+import Home from '../src/toolpad/Home';
 
 const Index: NextPage = () => <Home />;
 export default Index;
