+++ packages/toolpad-app/src/toolpad/AppEditor/PageEditor/PageModuleEditor.tsx
@@ -8,7 +8,7 @@ import { tryFormat } from '../../../utils/prettier';
 import useShortcut from '../../../utils/useShortcut';
 import lazyComponent from '../../../utils/lazyComponent';
 
-const TypescriptEditor = lazyComponent(() => import('../../TypescriptEditor'), {
+const TypescriptEditor = lazyComponent(() => import('../../../components/TypescriptEditor'), {
   noSsr: true,
   fallback: <Skeleton variant="rectangular" height="100%" />,
 });
