+++ packages/toolpad-app/src/toolpad/index.ts
@@ -1,0 +1,2 @@
+export * from './Toolpad';
+export { default } from './Toolpad';
